refactor: 이미지 파일 네이밍 변경
- SEO 정책에서 더 좋은 점수를 받기 위해 hypen 사용하여 이름 설정
