+++ frontend/src/components/common/Checkbox/Checkbox.tsx
@@ -2,7 +2,7 @@ import React, { InputHTMLAttributes, PropsWithChildren } from 'react';
 import styled from '@emotion/styled';
 import FlexContainer from '../FlexContainer/FlexContainer';
 import Check from '../../../assets/check.svg';
-import NotCheck from '../../../assets/not_check.svg';
+import NotCheck from '../../../assets/not-check.svg';
 
 interface Props extends PropsWithChildren<InputHTMLAttributes<HTMLInputElement>> {}
