+++ frontend/src/components/common/Radio/Radio.tsx
@@ -2,7 +2,7 @@ import styled from '@emotion/styled';
 import React, { InputHTMLAttributes, PropsWithChildren } from 'react';
 import FlexContainer from '../FlexContainer/FlexContainer';
 import Check from '../../../assets/check.svg';
-import NotCheck from '../../../assets/not_check.svg';
+import NotCheck from '../../../assets/not-check.svg';
 
 interface Props extends PropsWithChildren<InputHTMLAttributes<HTMLInputElement>> {
   name: string;
